Excel Import & Timesheet 2
I added my DYMO folder which contains a very basic set-up for importing
an excel spreadsheet. Currently, it's formatted to work with the fake
data spreadsheet which I included in the debug area in the folder.

IOExcel is the class that does the busy work, but Main needs to have
[STAThread] to function correctly.