Importing Excel and Modifying Database
I added the IOExcel class and MySQLClient class to the Registration
project. I added a few functions here and there to help with importing
and exporting to the online database.

By selecting the import button in the Admin tool, you should be able to
select an excel file to upload to the database. The Excel files are
formatted and named a certain way, but I see them more as a save file
that isn't meant to be tampered with anyway. The test.xlsx that
hopefully got included this time is an example of how it should be
formatted and can be uploaded to the database.

MySQLClient now contains methods for creating events and dropping them.