DYMO Printer functionality added
This update should make printing functional. There is one prior thing
that has to be done first before it will work though.

I included the jobfair.label file in the main area, but it needs to be
moved to the RegistrationKiosk/RegistrationKiosk/bin/Debug folder first
in order for it to work, since that's where the program is currently
being executed.