Label and Printer Fix / Timesheet
Note: I couldn't do any testing with student, since I can't seem to
register a student due it not accepting any student ID I put in.

ID Tags should now print as Shannon wants, varying on which Registered
Type they are. It should also be able to tell if the printer is
currently plugged in or not.