CSS/Pre-Registration Dynamic Text/Time Sheet
I added a .css file that makes the Pre-Registration form look a little
nicer.

Testing on the printer shows that all the text scales if one of the
names is too long. The default font size should be fine for most names,
though further testing on that one should still be done.

I also tampered with the preRegistrationForm file to handle dynamic
text, though it currently takes text from a local file.