blank sheet import bug - fix
For some reason, Excel was thinking that there was 100k+ rows to look
through, even though there was only 1. I set it to check the first cell
in each row, to see if it's blank. As soon as it finds a row that the
first cell is blank, it assumes that's the end. Not the best fix, but
it'll work.